code/requirements.txt
torch
numpy
Pillow
flask
requests
tqdm
diffusers
replicate
